+++ NEWS.md
@@ -53,6 +53,7 @@
 - Add a new `next_stage()` function to report the targets to be made in the next parallelizable stage.
 - Add a new `session_info` argument to `make()`. Apparently, `sessionInfo()` is a bottleneck for small `make()`s, so there is now an option to suppress it. This is mostly for the sake of speeding up unit tests.
 - Add a new `log_progress` argument to `make()` to suppress progress logging. This increases storage efficiency and speeds some projects up a tiny bit.
+- Add an optional `namespace` argument to `loadd()` and `readd()`. You can now load and read from non-default `storr` namespaces.
 
 # Changes in release 4.4.0